Fix for issue with justl-pop-to-buffer-on-display
When running 'justl' while `justl-pop-to-buffer-on-display` was set to nil, a blank buffer would be displayed.

This was caused by a misunderstanding of the purpose of `justl--pop-to-buffer`; the new option is only intended to apply to compilation/output buffers, rather than interaction buffers, and `justl--pop-to-buffer` is only used for interaction buffers.

The fix is to simply revert the behaviour of `justl--pop-to-buffer`, so it is unaffected by the new option.

Also updated the option description and the changelog to be more explicit about which buffers the option is for.